correction 100

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  */
 int main(void)
 {
-	long long x, y;
+	long int x, y;
 
 	x = 612852475143;
